feat: change not-found-search to server component
